add move new items actions and add for units
